Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,10 +2,10 @@
 
 int main()
 {
-    int a,b,c;
+    float a,b,c;
     float avg;
-    scanf("%d",&a);
-    scanf("%d",&b);
-    scanf("%d",&c);
-    printf("Average: %f",(a+b+c)/3);
+    scanf("%f",&a);
+    scanf("%f",&b);
+    scanf("%f",&c);
+    printf("Average: %.2f",(a+b+c)/3);
 }
